minor changes to api
